pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "mimic_videos"
version = "0.1.0"
description = "Video-Action Policy: generalist robot control via video diffusion priors"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "torch>=2.3",
    "torchvision",
    "diffusers>=0.32",
    "transformers>=4.45",
    "peft>=0.12",
    "einops>=0.8",
    "omegaconf>=2.3",
    "pyyaml",
    "wandb",
    "lerobot",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
include = ["mimic_videos*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
